% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_ROC.R
\name{calcAUC}
\alias{calcAUC}
\title{Calculates the area under the curve}
\usage{
calcAUC(roc_results, plot = TRUE)
}
\arguments{
\item{roc_results}{The results from the function \code{YARFROC}.}

\item{plot}{Make a plot of the ROC with the AUC shaded in? Default is \code{TRUE}.}
}
\value{
The area under the curve approximated via numerical integration.
}
\description{
Calculates the area under the curve
}
