% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_node_info.R
\name{compute_raw_proximity_info}
\alias{compute_raw_proximity_info}
\title{Computes information about the "proximity" of observations within the YARF model. Given two datasets,
information is computed for all pairs of observations. Information returned is the prediction nodes (for all trees) 
for both objects and their common node in the tree structure plus much information about the nodes.}
\usage{
compute_raw_proximity_info(
  yarf_mod,
  X1 = NULL,
  X2 = NULL,
  oob_only_1 = TRUE,
  oob_only_2 = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{yarf_mod}{A YARF model object.}

\item{X1}{A n* x p matrix where each row is an observation. If \code{NULL} (the default), 
then this parameter is set to the model's training data.}

\item{X2}{A n x p matrix where each row is an observation. If \code{NULL} (the default), 
then this parameter is set to the model's training data.}

\item{oob_only_1}{\code{TRUE} (the default) 
will only return proximity information for trees where each observation is out-of-bag. 
If \code{X1} is not \code{NULL}, this parameter has no effect on output.}

\item{oob_only_2}{\code{TRUE} (the default) 
will only return proximity information for trees where each observation is out-of-bag. 
If \code{X2} is not \code{NULL}, this parameter has no effect on output.}

\item{verbose}{Print out periodic messages on progress of the computation of the proximity information.
The default is \code{TRUE}.}
}
\value{
A list indexed by the row number of \code{X1} whose elements are a list indexed by
									the row number of \code{X2} whose elements are a list indexed by tree number whose
									elements are a list consisting of the following information: (a) X1 prediction node 
									(b) X2 prediction node, (c) shared node, (d) locations of all three nodes indicating
									left/right directions, (e) depth differences to shared node where 0 indicates they
									are in the same node, (f) parent node's cost for all three (where the parent cost
									of the root is the null model cost)
}
\description{
Computes information about the "proximity" of observations within the YARF model. Given two datasets,
information is computed for all pairs of observations. Information returned is the prediction nodes (for all trees) 
for both objects and their common node in the tree structure plus much information about the nodes.
}
\author{
Adam Kapelner
}
