% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_queries.R
\name{first_order_interaction_investigator}
\alias{first_order_interaction_investigator}
\title{This function investigates all two-way interactions in the forest by measuring how many times splits of one
variable is nested in splits of another variable.}
\usage{
first_order_interaction_investigator(yarf_mod, plot = TRUE,
  num_var_plot = 50, cut_bottom_prop = NULL, bottom_margin = 10, ...)
}
\arguments{
\item{yarf_mod}{The YARF model in which to query}

\item{plot}{This controls whether or not the function plots the interaction strengths. If so, it plots 
in order of strength from highest to lowest. Default is \code{TRUE}.}

\item{num_var_plot}{How many first order interactions to plot? All would be p-choose-2 which
can be large. The default is \code{50}. Use \code{Inf} to plot all p-choose-2.}

\item{cut_bottom_prop}{Cut the bottom of the barplot off at a percent of the minimum interaction strength. Default is
\code{NULL} for no cutting.}

\item{bottom_margin}{Allow for larger bottom margins. This is useful when the variables have long names.}

\item{...}{Additional parameters to be passed into the \code{barplot} function.}
}
\value{
A list with two components: (1) \code{interaction_counts} is a p x p sized matrix
							of interaction counts between the variables and (2) \code{sorted_counts} is a vector of
							the top interactions in order. Returns invisibly.
}
\description{
This function investigates all two-way interactions in the forest by measuring how many times splits of one
variable is nested in splits of another variable.
}
\author{
Adam Kapelner
}

