% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_queries.R
\name{get_tree_num_nodes_leaves_max_depths}
\alias{get_tree_num_nodes_leaves_max_depths}
\title{Gets information about each tree: the number of nodes, number of leaves and maximum depth.}
\usage{
get_tree_num_nodes_leaves_max_depths(yarf_mod)
}
\arguments{
\item{yarf_mod}{The YARF model in which to query}
}
\value{
A list with three integer vector components all of which have length equal to the 
						number of trees completes: (1) \code{num_nodes} has entries equal to the number of nodes
						in each tree, (2) \code{num_leaves} has entries equal to the number of leaves (terminal
						nodes) in each tree and (3) \code{max_depths} has entries equal to the maximum depth of
						each tree.
}
\description{
Gets information about each tree: the number of nodes, number of leaves and maximum depth.
}
\author{
Adam Kapelner
}
