% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_imaging.R
\name{illustrate_trees}
\alias{illustrate_trees}
\title{Illustrates trees to PNG files in the local directory.}
\usage{
illustrate_trees(
  yarf_mod,
  trees = NULL,
  max_depth = -1,
  background_rgb_color = c(255, 255, 255),
  line_rgb_color = c(100, 100, 100),
  text_rgb_color = c(0, 0, 0),
  font_family = "Arial",
  font_size = 12,
  margin_in_px = 70,
  character_width_in_px = 4.2,
  length_in_px_per_half_split = 20,
  depth_in_px_per_split = 100,
  print_at_split_node_script = NULL,
  print_at_leaf_script = NULL,
  file_format = "png",
  use_real_names = TRUE,
  title = "yarf_mod_tree",
  open_file = FALSE
)
}
\arguments{
\item{yarf_mod}{The YARF model whose trees are to be illustrated}

\item{trees}{A vector of the tree indices (a subset of 1...num_trees)}

\item{max_depth}{Illustrate the tree only up to this depth. Default is \code{-1}
indicating no restriction.}

\item{background_rgb_color}{A vector with 3 elements in \code{0, 1, ..., 255} specifying the RGB color
of the background. Default is \code{c(255, 255, 255)} for white.}

\item{line_rgb_color}{A vector with 3 elements in \code{0, 1, ..., 255} specifying the RGB color
of the lines. Default is \code{c(100, 100, 100)} for grey.}

\item{text_rgb_color}{A vector with 3 elements in \code{0, 1, ..., 255} specifying the RGB color
of the text. Default is \code{c(0, 0, 0)} for black.}

\item{font_family}{The font family of the text in the illustration. This string is passed in
as the first argument in a new \code{java.awt.Font} object. You can find
a list of legal arguments in the Java 8 documentation. Default is \code{arial}.}

\item{font_size}{The font size in the illustrations in pixels. Default is \code{12}.}

\item{margin_in_px}{The margin in pixels of the entire illustration (equal on all sides). Default
is \code{70}.}

\item{character_width_in_px}{Estimate as to the size of the character width in pixels. Differs based on fonts.}

\item{length_in_px_per_half_split}{The length of half of a split in pixels. Default is \code{20}.}

\item{depth_in_px_per_split}{The length of the depth of a split in pixels. Default is \code{100}.}

\item{print_at_split_node_script}{A custom javascript function which creates a special custom message to be printed
                                        at split nodes (see below). The default is \code{NULL} which prints no special message.

                                        function printAtSplitNode(node)\{ //node is of type YARF.YARFNode

                                            ...

                                            return message; //a string

                                        \}}

\item{print_at_leaf_script}{A custom javascript function which creates a special custom message to be printed
                                        at leaf nodes (see below). The default is \code{NULL} which prints no special message.

                                        function printAtLeaf(node)\{ //node is of type YARF.YARFNode

                                            ...

                                            return message; //a string

                                        \}}

\item{file_format}{The image format of the resulting illustration. Legal values are "jpg", "gif", "png". Default is "png".}

\item{use_real_names}{In the illustrations of split rules, set this to \code{TRUE} to use the names in the data frame.
and \code{FALSE} to use "X_1", "X_2", etc. Default is \code{TRUE}.}

\item{title}{The name of the file. Note that "_00t.png" will be added where "t"
is the tree index and "00" is an appropriate number of leading zeroes.
Default is "yarf_mod_tree".}

\item{open_file}{Should the first file be opened by the operating system? Default is \code{FALSE}.}
}
\description{
Illustrates trees to PNG files in the local directory.
}
\author{
Adam Kapelner
}
