% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_hypothesis_tests.R
\name{model_fit_test}
\alias{model_fit_test}
\title{Convenience method for testing a model's fit. The strategy is simple: one builds a YARF model from the 
residuals of the model you wish to fit and performs an omnibus test of all covariates against those 
residuals. This effecticely answers the question: "is there out-of-sample predictive information left over after 
this model was fit?" The p-val is determined by a permutation-like test.}
\usage{
model_fit_test(X, y, model_fit = NULL, num_permutation_samples = 100,
  plot = TRUE, ...)
}
\arguments{
\item{X}{The training data frame}

\item{y}{The vector of training responses}

\item{num_permutation_samples}{The resolution of the test. See \link{cov_importance_test}. The 
default is \code{100}.}

\item{plot}{Plot the result as a histogram. Default is \code{TRUE}.}

\item{...}{Additional parameters to be passed to \link{YARF}.}

\item{lin_mod}{An already-fit model that you wish to test if there
is any predictive power left over. The default is \code{NULL}
which indicates the standard linear model. Missingness not allowed.}
}
\value{
The same as \link{cov_importance_test}.
}
\description{
Convenience method for testing a model's fit. The strategy is simple: one builds a YARF model from the 
residuals of the model you wish to fit and performs an omnibus test of all covariates against those 
residuals. This effecticely answers the question: "is there out-of-sample predictive information left over after 
this model was fit?" The p-val is determined by a permutation-like test.
}
\author{
Adam Kapelner
}

