% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_predicts.R
\name{predict.YARF}
\alias{predict.YARF}
\title{Computes a response prediction for new data observation(s). Predictions are an aggregation from each of the 
individual trees' predictions based on the function passed into the model creation step. If no function was 
passed in, the default aggregation is the sample average for regression and the modal class for classification.}
\usage{
\method{predict}{YARF}(object, new_data, ...)
}
\arguments{
\item{object}{A YARF model object.}

\item{new_data}{A n* x p matrix where each row is an observation you wish to predict for. Here,
p is the same length as the original training data.}

\item{...}{Other methods passed to predict}
}
\value{
The yhat prediction for the \code{new_data} record(s).
}
\description{
Computes a response prediction for new data observation(s). Predictions are an aggregation from each of the 
individual trees' predictions based on the function passed into the model creation step. If no function was 
passed in, the default aggregation is the sample average for regression and the modal class for classification.
}
\author{
Kapelner
}
