% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_prints.R
\name{print.YARF}
\alias{print.YARF}
\title{Alias for summary.}
\usage{
\method{print}{YARF}(x, ...)
}
\arguments{
\item{x}{A YARF model object.}

\item{...}{Other methods passed to print}
}
\description{
Alias for summary.
}
\author{
Adam Kapelner
}

