% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_node_info.R
\name{proximity_info}
\alias{proximity_info}
\title{Computes information about the "proximity" of observations within the YARF model. Given two datasets,
information is computed for all pairs of observation.}
\usage{
proximity_info(yarf_mod, X, prox_single_node_calc_script = NULL)
}
\arguments{
\item{yarf_mod}{A YARF model object.}

\item{X}{A n* x p matrix where each row is an observation. If \code{NULL} (the default), 
then this parameter is set to the model's training data.}

\item{prox_single_node_calc_script}{Script ...}
}
\value{
(For now, a list of raw information for rows in each dataset)
}
\description{
Computes information about the "proximity" of observations within the YARF model. Given two datasets,
information is computed for all pairs of observation.
}
\author{
Matt Olson
}
