% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_queries.R
\name{query_variable_counts}
\alias{query_variable_counts}
\title{Function which queries the YARF model about how many times the variables are used when splitting}
\usage{
query_variable_counts(yarf_mod, type = "splits")
}
\arguments{
\item{yarf_mod}{The model in which to query}

\item{type}{Either "\code{trees}" which counts the number of trees where each variable was used one (or more) times
or "\code{splits}" which counts the number of times each variable was used across the whole model.}
}
\value{
A vector with length \code{p}, the number of variables considered in this YARF model, with either a binary
					value for \code{type = "trees"} or a count value for \code{type = "splits"}.
}
\description{
Function which queries the YARF model about how many times the variables are used when splitting
}
\author{
Adam Kapelner
}

