% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_queries.R
\name{query_variable_proportions_across_trees}
\alias{query_variable_proportions_across_trees}
\title{Function which queries the YARF model about the proportion of times the variables are used when splitting.
This is a convenience method which merely normalizes the result of \code{query_variable_counts}.}
\usage{
query_variable_proportions_across_trees(yarf_mod, type = "splits")
}
\arguments{
\item{yarf_mod}{The model in which to query}

\item{type}{Either "\code{trees}" which computes the proportion of trees where each variable was used one (or more) times
or "\code{splits}" which computes the proportion of times each variable was used across the whole model.}
}
\value{
A vector with length \code{p}, the number of variables considered in this YARF model, with values between
					0 and 1.
}
\description{
Function which queries the YARF model about the proportion of times the variables are used when splitting.
This is a convenience method which merely normalizes the result of \code{query_variable_counts}.
}
\author{
Adam Kapelner
}

