% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_node_info.R
\name{raw_node_data}
\alias{raw_node_data}
\title{Extract Raw Node Data}
\usage{
raw_node_data(yarf_mod, t = 1, location = "")
}
\arguments{
\item{yarf_mod}{A YARF model object.}

\item{t}{The tree to extract in {1, ..., num_trees} where the default is tree number 1.}

\item{location}{The location of the node to extract from indicated by a string consisting of "L"'s and "R"'s with no
spaces to indicate the traversal of the tree.}
}
\value{
\code{NULL} if the location does not exist otherwise a list object with the desired node's information
						where the nodes children are embedded in list objects within.
}
\description{
Extracts the structure of a tree as a list where the first element is the root node and the left/right daughters are
then embedded as lists and is recursed.
}
\author{
Adam Kapelner
}
