% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_builders.R
\name{set_YARF_num_cores}
\alias{set_YARF_num_cores}
\title{Sets the number of cores for YARF to use for all operations
which are multithreaded (e.g. model construction, prediction, etc)}
\usage{
set_YARF_num_cores(num_cores)
}
\arguments{
\item{num_cores}{The number of cores to use}
}
\description{
Sets the number of cores for YARF to use for all operations
which are multithreaded (e.g. model construction, prediction, etc)
}
\author{
Adam Kapelner
}
