% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_node_info.R
\name{shared_initial_substring}
\alias{shared_initial_substring}
\title{Returns the shared initial portion of a string}
\usage{
shared_initial_substring(s1, s2)
}
\arguments{
\item{s1}{The first string}

\item{s2}{The second string}
}
\value{
The shared substring
}
\description{
Given two strings, return the identical portion beginning from the first character and on.
}
\examples{
shared_initial_substring("money can't buy me love", "money can buy me love") 
#> "money can"
}
\author{
Adam Kapelner
}
