% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_prints.R
\name{summary.YARF}
\alias{summary.YARF}
\title{Prints a summary of the model to the console}
\usage{
\method{summary}{YARF}(object, ...)
}
\arguments{
\item{object}{A YARF model object.}

\item{...}{Other methods passed to summary}
}
\description{
Prints a summary of the model to the console
}
\author{
Adam Kapelner
}

