% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_node_info.R
\name{tree_average_proximity_info}
\alias{tree_average_proximity_info}
\title{Computes information about the "proximity" of observations within the YARF model. Given two datasets,
information is computed for all pairs of observations. Information is averaged over the trees in the model.}
\usage{
tree_average_proximity_info(raw)
}
\arguments{
\item{raw}{The information returned from \code{compute_raw_proximity_info()}.}
}
\value{
A list indexed by the row number of \code{X1} whose elements are a list indexed by
									the row number of \code{X2} whose elements are an array consisting of the following 
									information: (a) proportion of trees these two observations were in the same node,
									(b) average node 1/2/shared depth (c) average node 1/2/shared cost (d) average node 
									1/2/shared parent cost (e) average node 1/2/shared size (f) average depth difference 
									between node 1/2 and shared
}
\description{
Computes information about the "proximity" of observations within the YARF model. Given two datasets,
information is computed for all pairs of observations. Information is averaged over the trees in the model.
}
\author{
Kapelner
}
