% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF.R
\docType{package}
\name{YARF-package}
\alias{YARF-package}
\title{YARF: Yet Another Random Forests Package}
\description{
A customizable, asynchronous, and parallelized random forest implementation
with user-defined splitting and aggregation behavior, missingness incorporated
in attributes, imputation, variable importance, interaction detection, and
tree inspection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kapelner/YARF}
  \item \url{https://kapelner.github.io/YARF/}
  \item Report bugs at \url{https://github.com/kapelner/YARF/issues}
}

}
\author{
\strong{Maintainer}: Adam Kapelner \email{kapelner@qc.cuny.edu} (\href{https://orcid.org/0000-0001-5985-6792}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adam Kapelner \email{kapelner@qc.cuny.edu} (\href{https://orcid.org/0000-0001-5985-6792}{ORCID}) [copyright holder]
  \item Matt Olson
  \item Abhinav Patil
}

}
\keyword{internal}
