% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_oob_and_validation.R
\name{YARF_all_oob_results_matrix}
\alias{YARF_all_oob_results_matrix}
\title{All OOB Results Matrix}
\usage{
YARF_all_oob_results_matrix(yarf_mod)
}
\arguments{
\item{yarf_mod}{The yarf model object such that \code{YARF_update_with_oob_validation_results}
has been run before on a user-determined proportion of OOB samples 
as data validation.}
}
\description{
Computes the out-of-bag (OOB) predictions for all observations in training data by tree,
returning a matrix. This is useful for debugging or other analyses
}
\author{
Adam Kapelner
}
