% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_builders.R
\name{YARF_progress}
\alias{YARF_progress}
\title{Prints out a message reflecting the progress of the YARF model construction}
\usage{
YARF_progress(yarf_mod, console_message = TRUE)
}
\arguments{
\item{yarf_mod}{The yarf model object}

\item{console_message}{Should we print a message to console? Default is \code{TRUE}.}
}
\value{
The number of trees and the proportion completed.
}
\description{
Prints out a message reflecting the progress of the YARF model construction
}
\author{
Kapelner
}

