% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_builders.R
\name{YARF_progress_reports}
\alias{YARF_progress_reports}
\title{Prints out a messages reflecting the progress of the YARF model construction until completion}
\usage{
YARF_progress_reports(yarf_mod, time_delay_in_seconds = 10)
}
\arguments{
\item{yarf_mod}{The yarf model object}

\item{time_delay_in_seconds}{Frequency of messages in seconds. Default is \code{10} seconds.}
}
\description{
Prints out a messages reflecting the progress of the YARF model construction until completion
}
\author{
Kapelner
}

