% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_builders.R
\name{YARF_serialize}
\alias{YARF_serialize}
\title{Serializes the model so the user can use \code{save} and \code{save.image}
to write it to a file that can be then loaded into another and/or future R session.}
\usage{
YARF_serialize(yarf_mod)
}
\arguments{
\item{yarf_mod}{The yarf model object}
}
\description{
Serializes the model so the user can use \code{save} and \code{save.image}
to write it to a file that can be then loaded into another and/or future R session.
}
\author{
Adam Kapelner
}

