% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_validation.R
\name{YARF_update_with_oob_results}
\alias{YARF_update_with_oob_results}
\title{Computes the out-of-bag (OOB) predictions for the training data. 
This gives a good sense of out-of-sample performance in the future.}
\usage{
YARF_update_with_oob_results(yarf_mod, indices = NULL)
}
\arguments{
\item{yarf_mod}{The yarf model object}

\item{indices}{The indices to compute OOB performance. This is generally a private argument
for use by the validation / test functions.}
}
\description{
Computes the out-of-bag (OOB) predictions for the training data. 
This gives a good sense of out-of-sample performance in the future.
}
\author{
Adam Kapelner
}

