% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_validation.R
\name{YARF_update_with_oob_test_results}
\alias{YARF_update_with_oob_test_results}
\title{Computes the out-of-bag (OOB) predictions for the test subset of the training data. 
This function should only be run once. If it run again you risk adapting your model to the 
test data which will lead to overfitting and hence a biased out-of-sample performance estimate.}
\usage{
YARF_update_with_oob_test_results(yarf_mod)
}
\arguments{
\item{yarf_mod}{The yarf model object such that \code{YARF_update_with_oob_validation_results}
has been run before on a user-determined proportion of OOB samples 
as data validation.}
}
\description{
Computes the out-of-bag (OOB) predictions for the test subset of the training data. 
This function should only be run once. If it run again you risk adapting your model to the 
test data which will lead to overfitting and hence a biased out-of-sample performance estimate.
}
\author{
Adam Kapelner
}

