% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_validation.R
\name{YARF_update_with_oob_validation_results}
\alias{YARF_update_with_oob_validation_results}
\title{Computes the out-of-bag (OOB) predictions for the validation subset of the training data. 
This allows the user to try many different models while still leaving test data for 
true out-of-sample performance estimation.}
\usage{
YARF_update_with_oob_validation_results(yarf_mod, validation_proportion = 0.8)
}
\arguments{
\item{yarf_mod}{The yarf model object}

\item{validation_proportion}{The percentage of the data to use for validation data
(and the remaining is test data). The default is 80\%.}
}
\description{
Computes the out-of-bag (OOB) predictions for the validation subset of the training data. 
This allows the user to try many different models while still leaving test data for 
true out-of-sample performance estimation.
}
\author{
Adam Kapelner
}

