% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_imaging.R
\name{illustrate_trees}
\alias{illustrate_trees}
\title{Illustrates trees to PNG files in the local directory.}
\usage{
illustrate_trees(yarf_mod, trees = NULL, background_rgb_color = c(255, 255,
  255), line_rgb_color = c(100, 100, 100), text_rgb_color = c(0, 0, 0),
  font_family = "Arial", font_size = 12, margin_in_px = 70,
  character_width_in_px = 4.2, length_in_px_per_half_split = 20,
  depth_in_px_per_split = 100, title = "yarf_mod_tree")
}
\arguments{
\item{yarf_mod}{The YARF model whose trees are to be illustrated}

\item{trees}{A vector of the tree indices (a subset of 1...num_trees)}

\item{background_rgb_color}{A vector with 3 elements in {0, 1, ..., 255} specifying the RGB color 
of the background. Default is \code{c(255, 255, 255)} for white.}

\item{line_rgb_color}{A vector with 3 elements in {0, 1, ..., 255} specifying the RGB color 
of the lines. Default is \code{c(100, 100, 100)} for grey.}

\item{text_rgb_color}{A vector with 3 elements in {0, 1, ..., 255} specifying the RGB color 
of the text. Default is \code{c(0, 0, 0)} for black.}

\item{font_family}{The font family of the text in the illustration. This string is passed in 
as the first argument in a new \code{java.awt.Font} object. You can find 
a list of legal arguments in the Java 8 documentation. Default is \code{arial}.}

\item{font_size}{The font size in the illustrations in pixels. Default is \code{12}.}

\item{margin_in_px}{The margin in pixels of the entire illustration (equal on all sides). Default 
is \code{70}.}

\item{character_width_in_px}{Estimate as to the size of the character width in pixels. Differs based on fonts.}

\item{length_in_px_per_half_split}{The length of half of a split in pixels. Default is \code{20}.}

\item{depth_in_px_per_split}{The length of the depth of a split in pixels. Default is \code{100}.}

\item{title}{The name of the file. Note that "_00t.png" will be added where "t"
is the tree index. Default is "yarf_mod_tree".}
}
\description{
Illustrates trees to PNG files in the local directory.
}
\author{
Adam Kapelner
}

