% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_node_info.R
\name{prediction_nodes}
\alias{prediction_nodes}
\title{Returns the leaf nodes that are responsible for predictions for a given dataset.}
\usage{
prediction_nodes(yarf_mod, X = NULL, oob_only = TRUE)
}
\arguments{
\item{yarf_mod}{A YARF model object.}

\item{X}{A n* x p matrix where each row is an observation you wish to obtain prediction information. Here,
p is the same length as the original training data. If \code{NULL} (the default), 
then this parameter is set to the model's training data.}

\item{oob_only}{If \code{X} is \code{NULL}, setting this to \code{TRUE} (the default) will only return node information
for trees where each observation is out-of-bag. If \code{X} is not \code{NULL}, this
parameter has no effect on output.}
}
\value{
A list where values are rJava YARFNode objects for all the prediction nodes. The first key 
									indexes the row number in \code{X} and the second, the tree number.
}
\description{
Returns the leaf nodes that are responsible for predictions for a given dataset.
}
\author{
Kapelner
}
