% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF.R, R/YARF_builders.R
\docType{package}
\name{YARF}
\alias{YARF}
\alias{YARF-package}
\title{A fully customizable Random Forests implementation}
\usage{
YARF(X = NULL, y = NULL, Xy = NULL, Xother = NULL, num_trees = 500,
  boostrap_indices_fun = NULL, bootstrap_indicies_fun_data = NULL,
  oob_estimates = TRUE, mtry = NULL, cost_calc = NULL,
  node_assign = NULL, shared_functions = NULL, use_missing_data = FALSE,
  covariates_to_permute = NULL, use_missing_data_dummies_as_covars = FALSE,
  impute_missingness_with_x_j_bar_for_lm = TRUE, mem_cache_for_speed = TRUE,
  serialize = FALSE, seed = NULL, verbose = TRUE)
}
\arguments{
\item{X}{}

\item{y}{}

\item{Xy}{}

\item{Xother}{}

\item{num_trees}{}

\item{boostrap_indices_fun}{}

\item{bootstrap_indicies_fun_data}{}

\item{oob_estimates}{}

\item{mtry}{}

\item{cost_calc}{}

\item{node_assign}{}

\item{shared_functions}{}

\item{use_missing_data}{}

\item{covariates_to_permute}{}

\item{use_missing_data_dummies_as_covars}{}

\item{impute_missingness_with_x_j_bar_for_lm}{}

\item{mem_cache_for_speed}{}

\item{serialize}{}

\item{seed}{}

\item{verbose}{}
}
\value{
A list of all arguments passed in plus..
}
\description{
Software which helps practitioners optimally design experiments that measure the slope divided by the intercept.

Builds a YARF Model
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}

Adam Kapelner
}
\references{
TK
}
\keyword{Design,}
\keyword{Experimental}
\keyword{Slope}
\keyword{by}
\keyword{divided}
\keyword{intercept}

