% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_asynchronicity.R
\name{YARF_stop}
\alias{YARF_stop}
\title{Halts the model building.}
\usage{
YARF_stop(yarf_mod)
}
\arguments{
\item{yarf_mod}{The yarf model object}
}
\description{
Halts the model building.
}
\author{
Adam Kapelner
}
