% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_builders.R
\name{YARF_update_with_oob_results}
\alias{YARF_update_with_oob_results}
\title{Computes the out-of-bag (OOB) predictions for the training data. 
This gives a good sense of out-of-sample performance in the future.}
\usage{
YARF_update_with_oob_results(yarf_mod, oob_metric = NULL)
}
\arguments{
\item{yarf_mod}{The yarf model object}

\item{oob_metric}{If regression, we will return L1, MAE, L2 and RMSE and if classification,
we will return the confusion matrix with use/test errors and overall error rates. 
If you wish for another metric to be computed, pass the function in here as a string of 
javascript code. Default is \code{NULL} for no additional metric to be computed.}
}
\value{
The OOB results
}
\description{
Computes the out-of-bag (OOB) predictions for the training data. 
This gives a good sense of out-of-sample performance in the future.
}
\author{
Adam Kapelner
}

