% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YARF_ROC.R
\name{YARFROC}
\alias{YARFROC}
\title{ROC and ROC-like curves}
\usage{
YARFROC(X, y, x_axis = "FPR", y_axis = "TPR", use_prop_data = 1,
  minimum_class_proportion = 0.02, desired_interval = 0.05,
  tolerance = 0.001, y_axis_fine_resolution = FALSE, plot = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{X}{The covariate data frame}

\item{y}{The binary response vector}

\item{x_axis}{The rate to plot on the x-axis of the ROC like plot as a 3-character case-insensitive string. 
                        The default is "FPR" but it canbe any of the following:

                            "tpr" - true positive rate (AKA "sensitivity" or "recall") i.e the number predicted positive among those truly positive
                            "fnr" - false negative rate (AKA "miss rate") i.e the number predicted negative among those truly positive
                            "fpr" - false positive rate (AKA "fall-out") i.e. the number predicted positive among those truly negative
                            "tnr" - true negative rate (AKA "specificity") i.e. the number predicted negative among those truly negative
                            "ppv" - positive predictive value i.e. the number truly positive among those predicted positive 
                            "fdr" - false discovery rate i.e. the number truly negative among those predicted positive
                            "fomr" - false omission rate i.e. the number truly positive among those predicted negative
                            "npv" - negative predictive value i.e. the number truly negative among those predicted negative}

\item{y_axis}{The rate to plot on the y-axis of the ROC like plot as a 3-character string. 
The default is "TPR" but it can be any of those
listed in the documentation for the \code{x_axis} parameter. Must be different from \code{x-axis}.}

\item{...}{Additional arguments passed to YARF construction. This cannot include \code{bootstrap_indicies}.}
}
\value{
A data frame containing all
						models with the percent of first class in \code{levels(y)} for the bootstrap indices to replicate
						any point on the plot.
}
\description{
This generates data with a plot of an ROC curve or an ROC like curve for binary response
}
\author{
Adam
}
